chore(data-warehouse): move digest celery tasks into the product
