+++ products/data_warehouse/backend/external_data_source/test/test_tasks.py
@@ -1,0 +1,26 @@
+from unittest.mock import patch
+
+from products.data_warehouse.backend.tasks import (
+    send_external_data_failure_digest_catchup,
+    send_external_data_failure_digest_task,
+)
+
+
+class TestExternalDataFailureDigestTasks:
+    def test_digest_task_builds_digest(self):
+        with patch("products.data_warehouse.backend.tasks.notify_external_data_sync_failures") as mock_notify:
+            send_external_data_failure_digest_task(123)
+
+        mock_notify.assert_called_once_with(123)
+
+    def test_catchup_fans_out_per_team(self):
+        with (
+            patch(
+                "products.data_warehouse.backend.tasks.get_team_ids_with_recent_sync_failures",
+                return_value=[1, 2],
+            ),
+            patch("products.data_warehouse.backend.tasks.send_external_data_failure_digest_task") as mock_task,
+        ):
+            send_external_data_failure_digest_catchup()
+
+        assert [c.args for c in mock_task.delay.call_args_list] == [(1,), (2,)]
